Split browser integration tests from workspace model
