+++ docs/DECISIONS.md
@@ -259,3 +259,4 @@
 - Terminal command lifecycle mutation lives in `WorkspaceTerminalEngine`, not inline in `WorkspaceModel`. Input normalization, run-entry creation, streaming event application, missing execution-context failure, cancellation/stopped cleanup, marker cleanup, and completed-run cwd/environment persistence should stay in the terminal engine; `WorkspaceModel` owns selected-project sync, async shell streaming orchestration, top-bar updates, and user-facing errors.
 - Workspace command action planning and execution are separate boundaries. `WorkspaceCommandActionPlanner` maps command IDs and selected context to typed `WorkspaceCommandActionEffect` values, while `WorkspaceCommandActionExecutor` applies those effects through `QuillCodeWorkspaceModel` methods. `WorkspaceModel.runWorkspaceCommand` remains the command-plan entry point, but it should not own selected-state command action planning or the command-action effect switch.
 - Approval-card presentation keeps raw machine state separate from user-facing copy. Tool cards may remain `status == .review` for automation, tests, and action routing, while `ToolCardReviewState` carries the semantic review substate. Routine approvals display as `Ready` / `Ready to run` with `Run` and `Skip` actions. Only denied or policy-blocked review cards display `Needs review`. Native SwiftUI, static HTML, and the Playwright harness must read this copy from the tool-card surface instead of exposing raw safety jargon.
+- Browser integration coverage lives in `WorkspaceBrowserIntegrationTests`, not `WorkspaceModelTests`. Browser preview normalization, static/fetched snapshots, comments, history/reload, invalid-address errors, and composer-driven `host.browser.inspect` cross browser engine state, workspace surfaces, tool cards, and transcript events; keep those flows together while pure URL resolution and browser engine reducers stay in focused unit tests.
